chore: shut up clippy

+++ src/logger.rs
@@ -25,16 +25,15 @@ fn _setup() -> Result<(), fern::InitError> {
     };
     fs::create_dir_all(&path)?;
     let path = path.join("log.txt");
-    if let Ok(meta) = fs::metadata(&path) {
-        if meta.file_size() >= 5 * 1024 * 1024 {
+    if let Ok(meta) = fs::metadata(&path)
+        && meta.file_size() >= 5 * 1024 * 1024 {
             let _ = fs::remove_file(&path);
         }
-    }
     fern::Dispatch::new()
         .format(|out, message, record| {
             out.finish(format_args!(
                 "{} [{:<3}({})] {}",
-                Local::now().format("%Y-%m-%d %H:%M:%S").to_string(),
+                Local::now().format("%Y-%m-%d %H:%M:%S"),
                 record.level().abbr(),
                 ARCHITECHTURE,
                 message
